Extract client manager
