+++ backend/streams_explorer/streams_explorer.py
@@ -1,13 +1,12 @@
 import re
-from dataclasses import dataclass, field
 from typing import Dict, List, Optional, Type
 
 from cachetools.func import ttl_cache
 from fastapi import WebSocket
 from kubernetes_asyncio.client import V1beta1CronJob
 from loguru import logger
-from pydantic import BaseModel
 
+from streams_explorer.core.client_manager import ClientManager
 from streams_explorer.core.config import settings
 from streams_explorer.core.k8s_app import K8sApp
 from streams_explorer.core.node_info_extractor import (
@@ -34,33 +33,6 @@ from streams_explorer.models.node_information import (
     NodeInformation,
     NodeInfoType,
 )
-
-
-@dataclass
-class ClientManager:
-    _clients: List[WebSocket] = field(default_factory=list)
-
-    async def connect(self, websocket: WebSocket):
-        await websocket.accept()
-        logger.info("WebSocket client {} connected", websocket.client.host)
-        self._clients.append(websocket)
-
-    async def disconnect(self, websocket: WebSocket):
-        await websocket.close()
-        logger.info("WebSocket client {} disconnected", websocket.client.host)
-        self._clients.remove(websocket)
-
-    async def send(self, websocket: WebSocket, obj: BaseModel):
-        await websocket.send_json(obj.dict())
-
-    async def broadcast(self, obj: BaseModel):
-        data = obj.dict()
-        for client in self._clients:
-            await client.send_json(data)
-
-    @property
-    def clients(self) -> List[WebSocket]:
-        return self._clients
 
 
 class StreamsExplorer:
